+++ tests/unit_tests.py
@@ -403,8 +403,13 @@ def test_regex_parse():
     '''test date extraction using rules and regular expressions'''
     assert regex_parse_de('3. Dezember 2008') is not None
     assert regex_parse_de('33. Dezember 2008') is None
+    assert regex_parse_de('3 Aralık 2008 Çarşamba') is not None
+    assert regex_parse_de('3 Aralık 2008') is not None
     assert regex_parse_en('Tuesday, March 26th, 2019') is not None
+    assert regex_parse_en('March 26, 2019') is not None
     assert regex_parse_en('3rd Tuesday in March') is None
+    assert regex_parse_en('Mart 26, 2019') is not None
+    assert regex_parse_en('Salı, Mart 26, 2019') is not None
     assert regex_parse_en('3/14/2016') is not None
     assert regex_parse_en('36/14/2016') is None
     assert custom_parse('12122004', OUTPUTFORMAT, False, LATEST_POSSIBLE) is None
@@ -483,6 +488,18 @@ def test_search_html(original_date=False, max_date=LATEST_POSSIBLE):
     assert search_page('<html><body><p>© The Web Association 2013.</p></body></html>', OUTPUTFORMAT, original_date, max_date) == '2013-01-01'
     assert search_page('<html><body><p>Next © Copyright 2018</p></body></html>', OUTPUTFORMAT, original_date, max_date) == '2018-01-01'
 
+def test_idiosyncrasies():
+    assert find_date('<p><em>Last updated: 5/5/20</em></p>') == '2020-05-05'
+    assert find_date('<p><em>Published: 5/5/2020</em></p>') == '2020-05-05'
+    assert find_date('<p><em>Published in: 05.05.2020</em></p>') == '2020-05-05'
+    assert find_date('<p><em>Son güncelleme: 5/5/20</em></p>') == '2020-05-05'
+    assert find_date('<p><em>Son güncellenme: 5/5/2020</em></p>') == '2020-05-05'
+    assert find_date('<p><em>Yayımlama tarihi: 05.05.2020</em></p>') == '2020-05-05'
+    assert find_date('<p><em>Son güncelleme tarihi: 5/5/20</em></p>') == '2020-05-05'
+    assert find_date('<p><em>5/5/20 tarihinde güncellendi.</em></p>') == '2020-05-05'
+    assert find_date('''<p><em>5/5/20'de güncellendi.</em></p>''') == '2020-05-05'
+    assert find_date('<p><em>5/5/2020 tarihinde yayımlandı.</em></p>') == '2020-05-05'
+    assert find_date('<p><em>05.05.2020 tarihinde yayınlandı.</em></p>') == '2020-05-05'
 
 def test_parser():
     '''test argument parsing for the command-line interface'''
@@ -576,6 +593,7 @@ if __name__ == '__main__':
     test_search_html()
     test_url()
     test_approximate_url()
+    test_idiosyncrasies()
     # new_pages()
 
     # dependencies